refactor: separete requisitions types
