+++ src/services/app/payment/index.ts
@@ -1,1 +1,1 @@
-export { getAllPayments } from './payment';
+export { getAllPayments } from './payment.get';
